Restructure: made few restructurings to the project so that we can support scaling for backend functionality.
