+++ frontend/src/app/(protected)/settings/page.tsx
@@ -1,4 +1,4 @@
-import { auth, signOut } from "@/auth";
+import { auth, signOut } from "@/lib/auth";
 
 const SettingsPage = async () => {
   const session = await auth();
