+++ frontend/src/app/api/auth/[...nextauth]/route.ts
@@ -1,3 +1,3 @@
-export { GET, POST } from "@/auth"; // Referring to the auth.ts we just created
+export { GET, POST } from "@/lib/auth"; // Referring to the auth.ts we just created
 
 // export const { GET, POST } = handlers;
