+++ frontend/src/components/auth/register-form.tsx
@@ -19,7 +19,7 @@ import { Input } from "../ui/input";
 import { Button } from "../ui/button";
 import { FormError } from "../form-error";
 import { FormSuccess } from "../form-success";
-import { register } from "@/actions/register";
+import { register } from "@/utils/register";
 
 export const RegisterForm = () => {
   const [error, setError] = useState<string | undefined>("");
