+++ frontend/src/lib/auth.ts
@@ -1,5 +1,5 @@
 import NextAuth, { DefaultSession } from "next-auth";
-import authConfig from "@/auth.config";
+import authConfig from "@/utils/auth.config";
 import { getUserById } from "@/data/user";
 import { PrismaAdapter } from "@auth/prisma-adapter";
 import { db } from "@/lib/db";
