+++ frontend/src/middleware.ts
@@ -1,5 +1,5 @@
 import NextAuth from "next-auth";
-import authConfig from "@/auth.config";
+import authConfig from "@/utils/auth.config";
 
 import {
   DEFAULT_LOGIN_REDIRECT,
