+++ frontend/src/utils/auth.config.ts
