+++ frontend/src/utils/styret/fetchers.ts
