+++ frontend/src/utils/styret/uploadHS.ts
